isolate resizable component and refactor auctions
