+++ src/components/Futa/SearchableTicker/SearchableTicker.module.css
@@ -11,3 +11,173 @@
 .scrolling {
     scrollbar-color: var(--accent1) var(--dark2);
 }
+/* styling for the resizable table element */
+.resizable {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    justify-content: flex-start;
+    --width-for-scrollbar: 8px;
+    border: var(--futa-box-border);
+}
+
+/* Wrapper for ticker table rows and column headers */
+.ticker_table {
+    flex-direction: column;
+    width: 100%;
+    display: flex;
+    gap: 8px;
+    overflow-y: auto;
+    scrollbar-gutter: stable;
+}
+
+/* Header of the ticker table */
+.ticker_table > header {
+    position: sticky; /* Keeps header visible when scrolling */
+    top: 0;
+    display: flex;
+    flex-direction: row;
+    flex-wrap: nowrap;
+    justify-content: space-between;
+    gap: 8px;
+    background-color: var(--dark1);
+    color: var(--text2);
+    font-size: 12px;
+}
+
+/* Individual column headers in the ticker table */
+.ticker_table > header > * {
+    width: 120px;
+    flex: 1 10 auto;
+}
+
+/* Spacer to keep TICKER column header aligned with table data */
+.ticker_col_header_spacer {
+    visibility: hidden;
+}
+
+/* Container for ticker line items */
+.ticker_table_content {
+    display: flex;
+    flex-direction: column;
+    row-gap: 4px;
+}
+
+/* Left-aligned cell */
+.cell_left {
+    display: flex;
+    justify-content: flex-start;
+}
+
+/* Center-aligned cell */
+.cell_center {
+    display: flex;
+    justify-content: center;
+}
+
+/* Right-aligned cell */
+.cell_right {
+    display: flex;
+    justify-content: flex-end;
+}
+
+/* Differential content when length of data array is zero */
+.no_auctions_content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 8px;
+    padding: 2rem 0;
+    height: 100%;
+}
+
+.no_auctions_content > p {
+    color: var(--text2);
+}
+
+.no_auctions_content > button {
+    display: flex;
+    padding: 2px 8px;
+    justify-content: center;
+    align-items: center;
+    gap: 10px;
+    align-self: stretch;
+    outline: none;
+    background: var(--accent1, #0ccdff);
+    color: var(--dark1);
+    transition: all var(--animation-speed) ease-in-out;
+    margin: 0 auto;
+    text-transform: uppercase;
+}
+
+.no_auctions_content > button:hover {
+    background: var(--accent2);
+}
+
+/* Scrolling styles */
+.scrolling {
+    scrollbar-color: var(--accent1) var(--dark2);
+}
+
+/* Market cap and time remaining styles */
+.marketCap,
+.timeRemaining {
+    text-align: end;
+}
+
+/* Ticker name alignment */
+.ticker_name {
+    letter-spacing: 0.5px;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+
+/* Hover effect for ticker name arrow */
+.ticker_name_arrow {
+    padding-left: 4px;
+    opacity: 0;
+    transition: all var(--animation-speed) ease-in-out;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+}
+
+.tickerItemContainer:hover .ticker_name_arrow {
+    opacity: 1;
+}
+
+/* Status container styles */
+.statusContainer {
+    text-align: end;
+}
+
+.tickerStatus {
+    width: 10px;
+    height: 10px;
+    border-radius: 50%;
+}
+
+/* Responsive styles */
+@media (max-width: 500px) {
+    .ticker_table {
+        margin-left: 0px; /* Helps fit everything in the viewport */
+    }
+}
+
+@media (max-width: 400px) {
+    .ticker_table > header > * {
+        font-size: 10px;
+    }
+}
+
+@media (max-width: 1024px) {
+  
+    .resizable {
+        border: none;
+    }
+
+}
+
