+++ src/pages/platformFuta/Auctions/Auctions.module.css
@@ -1,3 +1,20 @@
+.resizable {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    justify-content: flex-start;
+    --width-for-scrollbar: 8px;
+    border: var(--futa-box-border);
+}
+.searchable_ticker {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    padding: 0 8px 8px 8px;
+    gap: 8px;
+}
+
 .mobileContainer {
     padding: 8px 16px;
     height: var(--content-height-mobile);
